+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project follows [Semantic Versioning](http://semver.org/).
 
 ## [11.3.0] 2026-06-11
 ### Added
+- Added form-level support for Questionnaire subjectType.
 - Added form-level support for Questionnaire meta.
 
 ## [11.2.3] 2026-06-10